Depend on both netty-internal and netty modules.
__Motivation__

It is better to be explicit about dependencies used in the module. This makes fixing dependencies between modules easy in terms of knowing which module is exactly required.
This module specifically was using transport-netty-internal as well as transport-netty module but was getting transport-netty transitively.

__Modification__

Specify explicit dependencies.

__Result__

Better management of dependencies.