Change BufferAllocators from an enum to static class
Motivation:
BufferAllocators is currently an enum. However it is not logically an enumeration and being an enumeration makes usage slightly more complicated by having to go through an accessor method to get the actually BufferAllocator. We can simplify the usage pattern if BufferAllocators was a static class with public static final members for each BufferAllocator.

Modifications:
- BufferAllocators should be a static class

Result:
Accessing the default BufferAllocator objects involves less layers.