Remove Http[Request|Response]Factory interfaces
Motivation:
We currently have HttpRequestFactory and HttpResponseFactory classes which allow the HTTP decoder which lives in the http-netty package to create HttpMetaData objects whose implementations live in the http-api module. However the intention is not to allow folks to customize how the HttpMetaData objects are created, or even wrap these objects to provide custom behavior in the decoder, but instead to control how headers/trailers are created. We can simplify the construction pattern and remove some variables if we remove the overridable factory interfaces for standalone static factories.

Modifications:
- Remove HttpRequestFactory and HttpResponseFactory
- Introduce HttpRequestMetaDataFactory and HttpResponseMetaDataFactory which are static factory classes with a single method

Result:
Simpler object creation for the http decoder classes.