Http headers factory
Motivation:
HttpHeaders and HttpTrailers are often created in the same place and with similar characteristics. We can consolidate the HttpHeadersFactory and the HttpTrailersFactory into a single interface to simplify the API and reduce the number of references to keep track of.

Modifications:
- Move HttpTrailersFactory methods into HttpHeadersFactory, and remove HttpTrailersFactory

Result:
One less interface to implement and keep track of.