Adding `executor` to `start` on `HttpServerStarter`.
Motivation:

When starting multiple services (on different sockets) from the same
`HttpServerStarter`, users will likely want to specify different
executors. We do not want to share them between servers. We also do not
want the `HttpServerStarter` to own the executors, which would require
it to close them.

Modifications:

Changing `executor` from a setter in the netty config, to a required
argument of `start`.

Results:

Users must specify an `executor`, there is no provided default. Users
may choose to share one between multiple servers, or not. Users are
also responsible for closing the executor.