Add option to disable checking for RefCountedTrapper.
Motivation:

When starting the HTTP server, we know we won't leak any ref counted
objects because the decoder takes care of them for us. We don't want
NettyConnectionContext to log a warning about the pipeline not containing
a RefCountedTrapper.

Modifications:

Added an optional boolean flag to NettyConnectionContext.newContext that
disables checking the pipeline for a RefCountedTrapper.

Results:

When the HTTP server is starting, it can specify not checking for a
RefCountedTrapper to avoid the unhelpful warning being logged.