HttpRequestDecoderTest#variableNoTrailers failure
Motivation:
variableNoTrailers test used the variable decoding (no encoding specified) but also wrote content with the request. For the request according to https://tools.ietf.org/html/rfc7230#section-3.3.3 number 6 this is invalid and should throw an error. The error was being thrown only when \r\n randomly appeared in the content, because the content was randomly generated. This resulted in a faulty unit test.

Modifications:
- The variableNoTrailers should not be writing content, and we should verify that the case of writing content is flagged as an error.

Result:
More reliable unit tests.