Fix redis tests
* Fix redis tests

__Motivation__

Redis tests are broken as they use the non-executor API to build the client.
This API uses a cached executor to offload signals which breaks the tests assumptions.
There is also an issue with flush: <rdar://problem/39874451> [ST-2.0] Using a non-immediate Executor for redis breaks flushes

__Modification__

Use `Executors.immediate()` for now. When we fix the other radar, we will add tests for non-immediate usage.

__Result__

Tests pass.