BlockingHttpClient
Motivation:
We have support for blocking APIs for HttpService but we do not have the
same support on the client side. These blocking APIs can be simpler to
consume for folks not accustomed to the Reactive Streams APIs.

Modifications:
- Introduce BlockingHttpRequester, BlockingHttpConnection,
BlockingHttpClient, and BlockingHttpClientGroup

Result:
HttpClient and related APIs have a blocking API variant.