ExecutionContext
Motivation:
Service filters have access to the ConnectionContext for each request, which provides access to the Executor, IoExecutor, and BufferAllocator. However Requester and Client APIs do not have access to the ConnectionContext because they may not be tied to a single connection. This makes filtering difficult on the client side without access to these items. We should provide access to these entities on the client side so creation of asynchronous sources and buffer allocation is possible.

Modifications:
- Move Executor, IoExecutor, and BufferAllocator into a ExecutionContext interface and have ConnectionContext extend this interface

Result:
Client side filters have access to the Executor and Buffer allocator.