ExecutionContext used during build time
Motivation:
Currently we require an IoExecutor, Executor, and BufferAllocator to
construct a client or a server. These items are configured in various
different places but may be required to be referenced together as an
ExecutionContext object as a portion of the ClientKey. Since this object
is used as a key we should keep it consistent and make it available to
the user for use in ClientKey, or just general use of each component.

Modifications:
- Builders should take an ExecutionContext instead of independently
requiring BufferAllocator, Executor, BufferAllocator
- Requester APIs should provide access to the ExecutionContext used
during build time.
- ConnectionContext should provide a ExecutionContext where the
IoExecutor is specific to the thread powering the IO for the
corresponding connection

Result:
More consistent API that allows folks to override/reuse the ClientKey
behavior.