HttpObjectDecoder cumulation index not always updated
Motivation:
HttpObjectDecoder keeps a cumulation index to ovoid revisiting input it
has already visited during the HTTP decoding process. This means
whenever the indexes change on the underlying cumulation buffer we must
update this index. However when data was read we were not updating this
index, and we were also setting the index to the writer index if we
failed to find a CRLF.

Modifications:
- Update the cumulationIndex when ever data is ready
- Don't force the cumulationIndex to writerIndex() if the CRLF wasn't
found and this was a larger value than we searched for
- Respect the maxHeaderSize, and maxInitialLineSize when looking for
CRLF

Result:
HttpObjectDecoder is able to parse larger chunks spread across multiple
packets.