Await varients which enforce non-null return type
Motivation:
Some of our examples and unit tests require an await method which enforce a non-null return value. However the existing await methods all permit null to be returned because null is a legit value for async sources with values.

Modifications:
- Add await*NonNull variants of all Await methods that may return null

Result:
Less boiler plate required for usage if null is not expected.