LoadBalancer event stream
Motivation:
LoadBalancer implementations often consume ServiceDiscovery events which
are delivered in an asynchronous fashion. This means that after a
LoadBalancer is created is is unlikely to be able to have connections
ready immediately. Users can apply a retry strategy with a busy loop, or
with a back off timeout approach but this is wasteful and may introduce
unnecessary latency. If the LoadBalancer could communicate when it is
ready to start processing requests we are much more likely to be
successful when attempting to issue requests.

Modifications:
- Introduce LoadBalancer#getEventStream that allows the LoadBalancer to
publish arbitrary events about its internal state
- Introduce a LoadBalancerReadyEvent which is a common event used by
load balancers to broadcast its readiness

Result:
LoadBalancer can communicate when it is ready to start processing
requests.