Account for load balancing state in the client and not the LoadBalancedConnection
Motivation:
The client implementations create a connection factory internally to avoid exposing the concrete type to external users. We currently manage some of the load balancer state in this concrete Connection, and some of it in the client. It is possible that someone will use the connection factory before the load balancer and therefore corrupt the load balancer state.

Modifications:
- The client should be responsible for updating all the load balancer state

Result:
Custom connection factories cannot corrupt the load balancer state inadvertently.