Client should use LoadBalancerReady filter by default
Motivation:
RedisClient is built synchronously but requests will fail until the LB is
ready. We have a LoadBalancerReadyRedisClient filter that is currently
not being used that can alleviate immediate request failures due to
initialization sequencing.

Modifications:

- The API to apply filters should provide access to the
LB#getEventStream() because the LB is built internally and the user has
no other way to access it.
- The default client filter should apply LoadBalancerReadyRedisClient
Simplify the Connection filter API from a Function to UnaryOperator

Result:
RedisClient can be used immediately without failures.