HttpObjectDecoder remove maxChunkSize
Motivation:
HttpObjectDecoder supports a maxChunkSize paramter which limits the size of buffers that are fired up the pipeline when reading the payload body. This behavior was inherited from Netty and it isn't clear that it provides any value for our use case. The memory for the data has already been committed so there is no sense in delaying its delivery or partitioning it across multiple objects.

Modifications:
- Remove the maxChunkSize limit from HttpObjectDecoder
- Remove the chunkedSupported flag from HttpObjectDecoder because we always want to support chunked encoding

Result:
Less object allocation required to read larger payload bodies.