Remove generics from HttpClient and HttpService
Motivation:
HttpClient and HttpService currently are generics types. The original
intention for the generic types was to help with filtering and data
transformations. On the client side this doesn't allow users to express
the type transformation information at the call site where the type
information is typically known in order to use the objects in business
logic. The filter would have to infer context from a disconnected
location in the code and maybe difficult to manage/reason about. On the
server side the generic types make it more challenging to convert
between all the different API variants (aggregated, blocking, etc...)
and also the type conversions are more typically a terminal event which
then directly interacts with application business logic. There are use
cases where preserving generics maybe useful but it is assumed this can
be built at higher layers.

Modifications:
- Remove the generic type arguments from HttpClient and HttpService

Result:
Simpler common case usage of HttpClient and HttpService, and more
intuitive type transformations on the HttpClient side.