ReadOnlyByteBuffer threading issues with constant values
Motivation:
ReadOnlyByteBuffer uses a cached temporary buffer in order to reduce object allocation for taking slices of the original buffer. However this causes issues when an instance is used by multiple threads. For example if the slice/duplicate is used and indexes of the underlying ByteBuffer are changed this may result in updating shared state between multiple threads in a non-thread safe manner and result in undefined behavior/exceptions.

Modifications:
- Avoid using a cached ByteBuffer when taking slices in ReadOnlyByteBuffer

Result:
ReadOnlyByteBuffer is now thread safe to read/consume across multiple threads.