Add unit test to verify log4j-provider.properties interface functions…
Motivation:
Log4j parses the log4j-provider.properties file as part of the static initialization but there is no way to verify if this was successful or not. This file also indicates an API version that may result in silently ignoring the provider during initialization.

Modifications:
- Add a unit test which verifies the provider is loaded

Result:
More test coverage for an implicit interface.