HttpClient hangs on HEAD request with Content-Length header in response
Motivation:
The HttpObjectDecoder and HttpObjectEncoder currently do not enforce the
RFC rules that require associating the request information with the
response information. For example if a client issues a HEAD request the
server may respond with a Content-Length header indicating how large the
response content would be. Since we don't associate request and response
information we currently don't know that no content should be expected
and hang.

Modifications:
- The Http encoder/decoder classes should associate the minimum amount
of state between request/response so that we can make progress for
exceptions in the RFC when no content type is allowed.

Result:
HttpClient no longer hangs when HEAD requests have Content-length header
in the response.