HttpHostHeaderFilter
Motivation:
When directly using a HttpClient or HttpConnection we currently do not
attempt to automatically infer the HOST header because the address is
unknown. This may lead to an invalid request from being sent with no
HOST header.

Modifications:
- HttpHostHeaderFilter which accepts a fallback address and can filter
HttpConnection and HttpClient objects.

Result:
We have the ability to automatically add a host header if not present.