Timeout operators and Completable#timer
Motivation:
 Applying a timeout to a processing chain is a common use case for which
we currently don't have support for.

Modifications:
- Add a timeout operator to Publisher, Single, and Completable
- Add a Completable#timer to provide a static source driven by timer
events that allows for compostability
- Modify Executor#schedule to return a Cancellable to be more consistent
with Executor#execute. This makes implementations of the timeout
operators simpler and require less object allocation. Folks can still
use the Completable#timer method to get the old behavior.
- NettyIoExecutor should have an asExecutor method so it can be used in
operators that require an Executor such as timeout. This is necessary if
folks have a threading model dependency on the Netty EventLoop threads.

Result:
Publisher, Single, Completable have a timeout operator.