RequestResponseCloseHandler updates
Motivation:
RequestResponseCloseHandler has some cases that result in failure to
close the connection and the state machine fails to transition to a
terminal state. It is also possible for the pending message count to
go negative which leads to undefined state.

Modifications:
- Fix cases where state machine doesn't terminate
- Make conditionals check pending first for short circuit and
consistency
- Fix Netty user event tests which are racy

Result:
More correct state machine in RequestResponseCloseHandler.