Introduce `CloseableIterable`
* Introduce `CloseableIterable`

__Motivation__

For deserialization, when we need to return an `Iterable`, we need to make sure that the underlying deserializer is disposed after the iterator is consumed.
Also, if the user does not want to consume any data, they can still dispose the deserializer. Since `Iterable` does not have an associated lifecycle, we need an `Iterable` that is closeable.

__Modification__

Added `CloseableIterable` and `CloseableIterator`.

__Result__

Well defined lifecycle for an `Iterable`.