Use ExecutionContext to start a server
Motivation:

The way we use to create and start server side is inconsistent with the client side, where we require an ExecutionContext to build the clients. Server side also need all these 3 objects: BufferAllocator, IoExecutor and Executor. But instead of ExecutionContext it requires them separately.
Moreover, it complicates resources clean up on the server side, when we create a default Executor internally. User should provide an ExecutionContext to start a server and we should never close any component of ExecutionContext internally.

Modifications:
- Require an `ExecutionContext` to start a new HTTP server;
- Remove all `start` methods, which provide default `Executor`;

Result:

Consistency in initialisation of client and server sides and their resources clean up.