Simplify retry/repeat strategies
* Simplify retry/repeat strategies

__Motivation__

`RetryStrategies` and `RepeatStrategies` were created before we had a way to schedule a timer using an `Executor`.
So, it required a `LongFunction` to schedule a timer. Now, an `Executor` can schedule a timer, so, we can simplify the APIs by just taking an `Executor`.

__Modification__

Modify the APIs to accept an `Executor` instead of a `LongFunction`.

__Result__

Simpler APIs.