AsyncClosable#closeAsyncGracefully with timeout as static method
Motivation:
AsyncClosable currently has two variations of closeAsyncGracefully. One with a timeout and one without a timeout. The one without the timeout is the only one that is currently actively used and facilitates composable usage. The variation with the timeout is not currently used but maybe useful in the future to ensure consistent behavior. The behavior of these two methods is not desired to be different and so instead of having both methods as members of the interface, the timeout variation can be made a static method to simplify the interface, avoid drift between the two methods, and serve as a utility method.

Modifications:
- AsyncClosable#closeAsyncGracefully with timeout as static method instead of a member of the interface

Result:
Simpler AsyncClosable interface where utility method role is clarified.