Avoid duplicating buffers during encoding
Motivation:
The current APIs for meta data related types return a Buffer and this is used by the encoder to copy that to a buffer which is then written to the wire. However the implementations of the buffer accessor APIs have to create a duplicate in order to avoid indexes being modified on the original buffer. This additional allocation creates unnecessary overhead demonstrated during wrk benchmarking and can be avoided via an alternative API.

Modifications:
- Instead of the meta data types returning a Buffer they can directly encode to a Buffer. This will avoid intermediate allocations and bounds checking.

Result:
Less overhead/allocation during encoding.