Remove HttpCookies
Motivation:
In order to modify the cookies in a HttpHeader you currently have to go
through HttpCookies. This can be expensive if the user facing API allows
for modifications of headers and cookies directly because each header
modifications has to check if the header name impacts the HttpCookies.
This can be avoided if individual headers can be created/encoded
independently from HttpCookies. There is a tradeoff made here that if
folks really need to do bulk HttpCookie manipulations they may incure
more encode/decode operations, but we can add the equivalent of
HttpCookies later as a utility on top of HttpHeaders if necessary.

Modifications:
- Add factory methods to HttpCookie to create new instances
- Add methods to HttpHeaders to get/add/remove HttpCookie objects
- Remove the HttpCookies interface

Result:
No longer necessary to keep state HttpCookies synchronized with
HttpHeaders in all cases.