Resolve ambiguity between ConnectionContext default methods overlapping with ExecutionContext
Motivation:
ConnectionContext has default methods that predate the addition of
ExecutionContext and eclipse the getExecutionContext() of this class. It
is not clear as to what the expected behavior would be if the
ConnectionContext actually did override these default methods, and even
worse if the implementation was inconsistent with the
getExecutionContext() values. This would lead to inconsistent usage and
undefined behavior.

Modifications:
- Remove the default methods from ConnectionContext that overlap with
the getExecutionContext() method

Result:
Less ambiguity around API usage and implementations for
ConnectionContext.