Publisher to Future and CompletionStage
Motivation:
We recently added support for Single to Future and CompletionStage, but not direct support for Publisher. If folks know the Publisher contents is already in memory, or it is known to be a finite stream and the user is willing to pay the memory and aggregation time, then these methods are useful.

Modifications:
- Add Publisher methods to convert toFuture() and toCompletionStage()
- Add Publisher methods that allow the user to control the aggregation strategy

Result:
Publisher can be easily converted to a Future and CompletionStage.