Completable to Future and CompletionStage
Motivation:
We recently added support for Single to Future and CompletionStage, but not direct support for Completable. Since Completable is often used in similar context as Single it allows the user to represent their semantics better than having to explicitly convert to Single first.

Modifications:
- Completable should support toFuture and toCompletionStage
- The toSingle and toPublisher conversion methods on the Completable API should support an option that provides no data, which is the default mode of operation for Completable

Result:
Completable supports Future and CompletionStage conversions. Completable to Publisher and to Single conversions are cleaner.