IterableMergeCompletable delayed error multiple terminate signals
Motivation:
IterableMergeCompletable supports delaying the propagation of an error
until all merged Completables terminate. In the case of merging an
Iterable we don't know how many total Completables will be merged so we
keep track when subscribe() is called and attempt to complete the merged
completable if they have all completed after the count is known. However
the implementation when delayed error was activated may have resulted in
trying to case the ON_COMPLETE object to a Throwable and terminate in
error if the Subscribers completed before the subscribe() call gets
around to attempting to complete.

Modifications:
- AbstractMergeCompletableOperator should avoid duplicate termination in
the event of delayedError, and use a DELIVERED_DELAYED_ERROR token
object to ensure single termination.

Result:
IterableMergeCompletable will terminate a single time and not throw a
ClassCastException.