HttpRequestEncoderTest#protocolPayloadEndOutboundShouldNotTriggerOnFailedFlush failure
Motivation:
HttpRequestEncoderTest#protocolPayloadEndOutboundShouldNotTriggerOnFailedFlush attempts to wait to write/flush until a RST event has been received so that the write operation will fail. However the current trigger point to allow the client to write is fired synchronously when the server initiates the close. This maybe before the server actual completes the close operation and before the client has processed the RST. This may lead to the write succeeding and the conditions the unit test was designed for not being verified.

Modifications:
- The client should wait until it has processed the reset before attempting to write/flush
- Instead of relying upon the flush strategy to delay the write/flush, we can just manually await the event before attempting the write

Result:
More reliable unit tests that exercises the client write fails scenario.