AbstractContextFilterChannelHandler state machine and visibility fixes
Motivation:
AbstractContextFilterChannelHandler's onContextFilterSuccessful maybe
executed asynchronously after a channelInactive event, but may result in
firing a channelActive event up the pipeline. This would violate Netty's
channel state machine and should be avoided.
AbstractContextFilterChannelHandler lazily creates a
SequentialCancellable before execution of the tasks which invokes the
user's context filter. This SequentialCancellable is then used in the
Subscriber#onSubscribe method. RS specifies that synchronization for
state much be provided externally and the executor may not provide any
synchronization in this case. If the subscribe is done on another thread
it the SequentialCancellable may not be visible.

Modifications:
- Add a CHANNEL_INACTIVE state to the state machine, ensure this is
a terminal state and that the onContextFilterSuccessful is not called if
this state is reached.
- Make the SequentialCancellable final for visibility

Result:
More correct state machine in AbstractContextFilterChannelHandler and
visibility of shared state.