Rename HTTP API variations
Motivation:
Initial feedback has demonstrated that the "aggregated" terminology is
confusing for such a prevalent API. The "streaming" APIs are currently
not qualified but assumed either to not be the most commonly used, or if
they are commonly used adding the "streaming" terminology will help more
clearly distinguish from the aggregated use cases.

Modifications:
- Drop the "aggregated" term from public APIs. These APIs will not be
qualified and assumed to be the "default" or first to be discovered
variants.
- Add the "streaming" term to APIs which support streaming. This maybe
asynchronous (which has no explicit qualifier) or blocking (which does
have a qualifier).
- Retain the "blocking" terminology in API names so we/users can more
easily identify blocking use cases if they attempt to partially migrate
toward async/streaming.

Result:
APIs that clarify their intended use case, and are adjusted for use
cases assumed to be more prevalent.