Override `FlushStrategy` per connection.
__Motivation__

We should provide a consistent way to specify custom `FlushStrategy`. Currently, HTTP does not have a way to do this.
Our current thinking is that when custom flush strategies are used, they are effective when writes are controlled, i.e. there are no concurrent writes and hence the user is "reserves" the connection either explicitly (client) or implicitly (server). Hence, if we provide this ability on a connection, it should be sufficient.

__Modification__

- Add a new internal API: `NettyConnectionContext`.
- `NettyConnectionContext` can be used to `updateFlushStrategy()` or to get `getConnectionEvents()`.

The above can be used on the server side by casting `ConnectionContext` to `NettyConnectionContext`. On the client, side it is only available when `ConnectionContext` is in scope.

__Result__

More flexibility of flushes.