Merge `HttpSerializerProvider` and `HttpDeserializerProvider`
__Motivation__

It is useful for users to have a single entity for both serialization and deserialization instead of carrying two entities where ever (de)serialization is required.
So, having serialization and deserialization in a single entity is useful.

__Modification__

Merged `HttpSerializerProvider` and `HttpDeserializerProvider` into `HttpSerializationProvider`
Also merged `HttpSerializerProviders` and `HttpDeserializerProviders` into `HttpSerializationProviders`

__Result__

Co-location of commonly used concerns.