Introducing RequestHandlers
__Motivation__

`*HttpService` interfaces are a bit verbose as they are abstract classes and contain lifecycle methods.
For simpler services that do not carry state, we have `from*` methods that take a lambda to wrap.
We can make the lambda as a service that `HttpServerStarter` can directly take which will simplify the simple cases without limiting the complex cases of having state in service.

__Modification__

Introduced a RequestHandler abstraction for all variants and have `HttpServerStarter` take these handlers instead.
Since the lifecycle aware services extends these handlers this means both handlers and services can be passed to the start methods.

__Result__

More discoverable path for simple services.