Never abort requests on client Socket Input Shutdown
__Motivation__

In case of a client connection the RequestResponseCloseHandler would abort requests prematurely on Socket Input Shutdown,
assuming the response couldn't be fully consumed. This allows the user to decide how to deal with this situation.

__Modifications__

Prevent the state machine from closing the connection until the current request has completed sending.

__Result__

User can now decide to interrupt the request in flight or let it carry on.