Exclude BOM modules from IDEA projects dependencies, Disable JerseyDataJacksonTestSuite if Jersey's JacksonFeature is on the classpath (#83)
## Motivation

The IntelliJ control files generated wrongly include bom projects on the classpath.

They also wrongly share the test classpath of `servicetalk-http-router-jersey` with `servicetalk-data-jackson-jersey` which unfortunately gives access to `jersey-media-json-jackson`.

## Modifications

- Filter `bom` dependencies from the generated IntelliJ project classpath.

Fixing the second issue involves generating project control files similar to the ones IntelliJ creates when a Gradle project is imported [1],
i.e. with different modules for different scopes. This allows bypassing this limitation:

> At runtime, all the dependencies of the current module are included in the classpath for the modules that depend on this module. [2]

This is a non-trivial amount of work that imvolves post-processing the `ipr` and `ipl` models laoded by the Gradle IDEA plugin to make them behave like imports in IntelliJ.
Therefore this PR instead:

- Disable `JerseyDataJacksonTestSuite` if Jersey's `JacksonFeature` is on the classpath, with a clear reason message.

[1] The Gradle IDEA plugin even recommends this way: https://docs.gradle.org/current/userguide/idea_plugin.html
[2] https://www.jetbrains.com/help/idea/dependencies.html

## Results

- No more `bom` dependencies for projects build with `servicetalk-gradle-plugin-internal`,
- No more mysterious failures when running the `JerseyDataJacksonTestSuite` in IntelliJ.