ClientFilterFactory consistency and OpenTracing filter cleanup
Motivation:
ClientFilterFuction is specific to Http, but doesn't indicate Http in
the type name. This is inconsistent with RedisClientFilterFunction. Also
there are missing Redis*FilterFunction APIs.
OpenTracing filter implementation added an additional method to
AsyncContext which isn't necessary with the current use cases.

Modifications:
- Rename ClientFilterFactory, ConnectionFilterFactory,
GroupedClientFilterFactory to have Http in the name
- Rename RedisClientFilterFunction to RedisClientFilterFactory
- Add missing Redis filter APIs
- Remove the AsyncContext#wrap method for Subscriber and remove its
usage in TcpConnector
- Reduce the number of operators in the OpenTracing filters by using
doAfterSubscriber

Result:
More consistency in filter factories, less methods in AsyncContext, less
operators in OpenTracing filters.