Selective offloading
__Motivation__

Today there is only one way to influence offloading done by client/server by specifying or omitting an `Executor` for that client/server.
This works well for homogeneous clients/server where there is just one execution model: either offload or not. However, in reality client/servers may need offloading for some execution paths but others can be asynchronous.
We should offer a way to override the offloading strategy per request/route, for which we should do the following:

- One should be able to specify the offloading strategy while making request to a client.
- One should be able to specify a different offloading strategy per route on a server. Since, we need the strategy to even call the `Service`, this strategy should be provided before we call the `Service`.
-- Our predicate router should be able to handle different routes with different offloading strategies.

__Modification__

Introduce the concept of `ExecutionStrategy` in place of `Executor` which can be specified at the following places:

- While making requests to the clients.
- By implementing a method `executionStrategy()` on a service.

`ExecutionStrategy` influences the following aspects of offloading:

- Which `Executor` should be used for offloading.
- Whether sending of data (request for client, response for server) should be offloaded.
- Whether receiving of data (response for client, request for server) should be offloaded.
- Whether calling the service should be offloaded.
- For HTTP: Whether invoking the `Single` returned by the client should be offloaded.

Static factories are provided per protocol to get a suitable strategy for use.

__Result__

Flexibility of offloading per request.