Fix synchronous resubscribe on NettyChannelPublisher
__Motivation__

In a scenario with multiple payloads or data and an error queued up in NettyChannelPublisher, when a new Subscriber synchronously re-subscribes and demands data from the terminal event, the draining loop exits early and ignores the demand. In the case of pipelined HTTP requests, this lead to the eagerly requested metadata never getting delivered.

__Modifications__

Make sure NettyChannelPublisher continues draining the queue in the event that the subscriber is changed as part of delivering the terminal event.

__Result__

Responses to a pipelined request don't appear to get stuck when queued up.