Service filters for Http server
__Motivation__

Adding filters for `HttpServerBuilder` is not consistent with client.
For server, users are required to manually wrap the service, for clients, we provide a method on the builder.
We should be providing similar method on the server.

__Modification__

Added `HttpServerBuilder#appendServiceFilter` to add filters.

__Result__

Easier way to add filters for server.