RedisClient#reserveConnection API simplification (#105)
Motivation:
RedisClient#reserveConnection currently takes a RedisRequest, but does not execute it. This API is flexible in that it allows the underlying RedisClient implementation to apply different policies about how the underlying connection is managed if pooling is used. However the API has caused confusion for users because they may use the resulting connection for multiple requests, and they also have to know the request isn't actually issued and call request(..) again.

Modifications:
- Remove the RedisRequest from RedisClient#reserveConnection and instead just take the Command. The Command is currently required because connection implementation behavior is different depending upon command type (e.g. subscribe).

Result:
Simplified RedisClient#reserveConnection with less chance to confuse the purpose of the RedisRequest.