RetryingHttpClientFilter#build method remove retryCount
Motivation:
RetryingHttpClientFilter#build requires a retryCount argument. In general putting arguments on the build() methods are most useful when they are likely to change between invocations, they are required, and a good default is not feasible. However in this case we can simplify the build() method and provide a default value for the number of retry counts.

Modifications:
- RetryingHttpClientFilter#build to not take any arguments
- Add a retryCount method on the builder and set the default value to 2

Result:
RetryingHttpClientFilter adheres to the more traditional builder pattern and provides a default value for retry count.