AsyncContextMap and PartitionAttributes Key factory method simplify (#123)
Motivation:
AsyncContextMap and PartitionAttributes have inner class Key which is compared by reference. However for debugging and visibility there is a factory method that allows a string to be provided. The method name is newKeyWithDebugToString to emphasize the String value will not be used in comparisons. However this is seen as overly verbose and the javadocs clarify the purpose adequately.

Modifications:
- rename newKeyWithDebugToString to newKey

Result:
Less verbose Key creation.